requirements.txt
# Core dependencies
beautifulsoup4>=4.12.0
certifi>=2023.0.0
charset-normalizer>=3.0.0
click>=8.0.0
colorama>=0.4.6
fire>=0.5.0
idna>=3.0
packaging>=23.0
requests>=2.28.0
rich>=13.0.0
six>=1.16.0
tqdm>=4.65.0
typing_extensions>=4.5.0
urllib3>=2.0.0

# Windows automation (Windows only)
comtypes>=1.2.0
pywin32>=305
pywinauto>=0.6.8
uiautomation>=1.1.0

# AI & ML
google-generativeai>=0.3.0
grpcio>=1.50.0
protobuf>=4.0.0

# Face detection (optional - comment out if not needed)
dlib-bin>=19.24.0
retina-face>=0.0.13

# Video download
yt-dlp>=2023.3.4

# Web scraping
selenium>=4.8.0
soupsieve>=2.4
websocket-client>=1.5.0
websockets>=11.0
wsproto>=1.2.0
trio>=0.22.0
trio-websocket>=0.10.0

# System utilities
psutil>=5.9.0
pyperclip>=1.8.0

# Misc
attrs>=23.0.0
filelock>=3.12.0
python-dateutil>=2.8.0
pytz>=2023.3
sortedcontainers>=2.4.0
